requirements.txt
asgiref==3.9.1
certifi==2025.8.3
chardet==5.2.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pypng==0.20220715.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qrcode==7.4.2
reportlab==4.2.2
sentry-sdk==2.34.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
